Projects/FairLift.py/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Cors==4.0.0
Werkzeug==3.0.3
PyMySQL==1.1.1
python-dotenv==1.0.1